reafactoring - move the package metadata logic in the devtools package
