+++ packages/sitecore-jss-dev-tools/src/templating/index.ts
@@ -2,3 +2,4 @@ export { ComponentFile, PackageDefinition, getComponentList } from './components
 export { PluginDefinition, generatePlugins, ModuleType } from './plugins';
 export { scaffoldFile } from './scaffold';
 export { strip } from './strip';
+export { getPackagesMetadata } from './metadata';
